Crate zng_wgt_rule_line

Rule line widgets.

Crate

This crate is part of the zng project.

Modules

hr

Horizontal rule line.

vr

Vertical rule line.

Structs

COLLAPSE_SCOPE

Contextual service managed by collapse_scope.

CollapseMode

Represents what rule lines are collapsed in a collapse_scope.

RuleLine

W Draws a horizontal or vertical rule line.

Statics

COLLAPSABLE_LINE_ID

Identifies a line widget that can be collapsed by collapse_scope.

COLLAPSE_SKIP_ID

Identifies a widget (and descendants) to be ignored by the collapse_scope.

Functions

collapse_scope

P Collapse adjacent descendant rule lines.

collapse_skip

P Defines if this widget and descendants are ignored by collapse_scope.

color

B Line color.

length

B Line length.

line_style

B Line style.

orientation

B Line orientation.

stroke_thickness

B Line stroke thickness.