zng_wgt_rule_line

Module hr

Horizontal rule line.

Structs

Hr

W Draws an horizontal RuleLine!.

Statics

COLOR_VAR

Line color, inherits from FONT_COLOR_VAR.

LINE_STYLE_VAR

Line style, default is Solid.

MARGIN_VAR

Margin around line.

STROKE_THICKNESS_VAR

Line stroke thickness, default is 1.dip()

WIDTH_VAR

Horizontal line length.

Functions

color

P Sets the line color of all descendant Hr!().

line_style

P Sets the line style of all descendant Hr!().

margin

P Sets the margin around line of all descendant Hr!().

stroke_thickness

P Sets the line stroke thickness of all descendant Hr!().

width

P Sets the horizontal line length of all descendant Hr!().