zng_wgt_rule_line

Module vr

Vertical rule line.

Structs

Vr

W Draws a vertical RuleLine!.

Statics

COLOR_VAR

Line color, inherits from FONT_COLOR_VAR.

LINE_STYLE_VAR

Line style, default is Solid.

MARGIN_VAR

Margin around line.

STROKE_THICKNESS_VAR

Line stroke thickness, default is 1.dip()

Functions

color

P Sets the COLOR_VAR that affects all vertical rules inside the widget.

line_style

P Sets the LINE_STYLE_VAR that affects all vertical rules inside the widget.

margin

P Sets the MARGIN_VAR that affects all vertical rules inside the widget.

stroke_thickness

P Sets the STROKE_THICKNESS_VAR that affects all vertical rules inside the widget.